engineeringindia-backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
psycopg2-binary>=2.9
python-dotenv>=1.0.0
requests>=2.28
dj-database-url>=1.0
PyJWT>=2.8
whitenoise
gunicorn
